:recycle: api: move certification-report-controller to src
