+++ api/src/certification/session/application/certification-report-route.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 import { certificationReportController } from './certification-report-controller.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
-import { authorization } from '../preHandlers/authorization.js';
-import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 
 const register = async function (server) {
   server.route([
