+++ api/tests/unit/application/certification-report/certification-report-controller_test.js
@@ -1,5 +1,5 @@
 import { sinon, expect, hFake, domainBuilder } from '../../../test-helper.js';
-import { certificationReportController } from '../../../../lib/application/certification-reports/certification-report-controller.js';
+import { certificationReportController } from '../../../../src/certification/session/application/certification-report-controller.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 
 describe('Unit | Controller | certification-report-controller', function () {
