+++ api/tests/unit/application/certification-report/index_test.js
@@ -1,5 +1,5 @@
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-import { certificationReportController } from '../../../../lib/application/certification-reports/certification-report-controller.js';
+import { certificationReportController } from '../../../../src/certification/session/application/certification-report-controller.js';
 import { NotFoundError } from '../../../../lib/application/http-errors.js';
 import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
